requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
python-dotenv>=0.19.0
httpx>=0.24.0
pyyaml>=6.0.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=2.0.0 